# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для сканирования портов и просмотра истории запросов",
        "title": "Network Scanner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/history": {
            "get": {
                "description": "Получить историю всех сканирований",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "История сканирований",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ScanRequest"
                            }
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/results/{id}": {
            "get": {
                "description": "Получить результаты по request_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "Результаты сканирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID запроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ScanResult"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scan": {
            "post": {
                "description": "Сканирует указанные порты на IP и сохраняет результат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "Сканировать порты",
                "parameters": [
                    {
                        "description": "Запрос на сканирование",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScanRequestSwagger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ScanResponseSwagger"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ScanRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_address": {
                    "type": "string"
                },
                "ports": {
                    "description": "Может быть \"80,443\" или \"1-1024\"",
                    "type": "string"
                }
            }
        },
        "models.ScanRequestSwagger": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "ports": {
                    "type": "string",
                    "example": "22,80,443"
                }
            }
        },
        "models.ScanResponseSwagger": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "open_ports": {
                    "description": "\u003c--- так",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        22,
                        80
                    ]
                },
                "ports": {
                    "type": "string",
                    "example": "22,80,443"
                },
                "request_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.ScanResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_open": {
                    "type": "boolean"
                },
                "port": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "integer"
                },
                "scanned_at": {
                    "type": "string"
                }
            }
        }
    }
}